+++ examples/common/BusDiag.h
@@ -1,7 +1,0 @@
-#pragma once
-
-#include "E2Diagnostics.h"
-
-namespace bus_diag {
-using namespace e2diag;
-}